refactor: 게시글 다중 삭제 비즈니스 로직에 구현한 함수 적용

+++ src/main/java/econo/buddybridge/post/event/handler/PostDeleteEventHandler.java
@@ -2,17 +2,18 @@ package econo.buddybridge.post.event.handler;
 
 import econo.buddybridge.chat.chatmessage.repository.ChatMessageRepository;
 import econo.buddybridge.chat.chatmessage.repository.MessageReadStatusRepository;
-import econo.buddybridge.comment.entity.Comment;
 import econo.buddybridge.comment.repository.CommentRepository;
 import econo.buddybridge.matching.entity.Matching;
 import econo.buddybridge.matching.repository.MatchingRepository;
 import econo.buddybridge.post.entity.Post;
-import econo.buddybridge.post.entity.PostLike;
 import econo.buddybridge.post.event.PostDeleteEvent;
 import econo.buddybridge.post.repository.PostLikeRepository;
 import econo.buddybridge.post.repository.PostRepository;
+import econo.buddybridge.report.repository.CommentReportRepository;
+import econo.buddybridge.report.repository.MatchingReportRepository;
 import econo.buddybridge.report.repository.PostReportRepository;
 import java.util.List;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
@@ -23,6 +24,9 @@ import org.springframework.transaction.event.TransactionalEventListener;
 public class PostDeleteEventHandler {
 
     private final PostReportRepository postReportRepository;
+    private final CommentReportRepository commentReportRepository;
+    private final MatchingReportRepository matchingReportRepository;
+
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
     private final MatchingRepository matchingRepository;
@@ -34,26 +38,33 @@ public class PostDeleteEventHandler {
     public void handlePostDeleteEvent(PostDeleteEvent event) {
         List<Post> posts = event.getPosts();
 
-        List<Post> reportedPosts = postReportRepository.findByReportedPostIn(posts);
+        Set<Post> reportedPosts = postReportRepository.findByReportedPostIn(posts);
+        Set<Post> reportedCommentPosts = commentReportRepository.findPostByReportedCommentPostIn(posts);
+        Set<Post> reportedMatchingPosts = matchingReportRepository.findPostByReportedMatchingPostIn(posts);
+
+        reportedPosts.addAll(reportedCommentPosts);
+        reportedPosts.addAll(reportedMatchingPosts);
+
         // 신고된 게시글은 soft delete
-        reportedPosts.forEach(Post::delete);
+        if (!reportedPosts.isEmpty()) {
+            postRepository.softDeleteAllIn(reportedPosts);
+        }
 
         // 신고된 게시글이 아닌 경우 완전 삭제
         posts = posts.stream()
                 .filter(post -> !reportedPosts.contains(post))
                 .toList();
 
-        List<Matching> matchings = matchingRepository.findByPostIn(posts);
-        chatMessageRepository.deleteAllInBatch(chatMessageRepository.findByMatchingIn(matchings));
-        messageReadStatusRepository.deleteAllInBatch(messageReadStatusRepository.findByMatchingIn(matchings));
-        matchingRepository.deleteAllInBatch(matchings);
+        if (!posts.isEmpty()) {
+            List<Matching> matchings = matchingRepository.findByPostIn(posts);
+            chatMessageRepository.deleteAllByMatchingIn(matchings);
+            messageReadStatusRepository.deleteAllByMatchingIn(matchings);
+            matchingRepository.deleteAllInBatch(matchings);
 
-        List<Comment> comments = commentRepository.findByPostIn(posts);
-        commentRepository.deleteAllInBatch(comments);
+            commentRepository.deleteAllByPostIn(posts);
+            postLikeRepository.deleteAllByPostIn(posts);
 
-        List<PostLike> postLikes = postLikeRepository.findByPostIn(posts);
-        postLikeRepository.deleteAllInBatch(postLikes);
-
-        postRepository.deleteAllInBatch(posts);
+            postRepository.deleteAllIn(posts);
+        }
     }
 }
